mysite/requirements.txt
Django
channels
channels_redis
asgi-redis
django-crispy-forms
django-utils-six
Pillow

